+++ modules/core/src/main/kotlin/io/github/lmliam/kotventure/core/selector/EntitySelectorAdapter.kt
@@ -13,4 +13,12 @@ internal class EntitySelectorAdapter(
     override fun type(entityType: String) {
         state.assignType(entityType)
     }
+
+    override fun typeTag(entityTypeTag: Key) {
+        state.assignTypeTag(entityTypeTag)
+    }
+
+    override fun not(init: NegatedEntitySelectorScope.() -> Unit) {
+        NegatedEntitySelectorAdapter(state).init()
+    }
 }