+++ modules/core/src/main/kotlin/io/github/lmliam/kotventure/core/selector/EntitySelectorScope.kt
@@ -8,6 +8,13 @@ import io.github.lmliam.kotventure.core.dsl.KotventureDslMarker
  * @sample io.github.lmliam.kotventure.core.selector.entitySelectorScopeSample
  */
 @KotventureDslMarker
-public interface EntitySelectorScope :
+public sealed interface EntitySelectorScope :
     PlayerEntitySelectorScope,
-    SelfEntitySelectorScope
+    SelfEntitySelectorScope {
+    /**
+     * Applies negated filters to this full entity selector.
+     *
+     * @sample io.github.lmliam.kotventure.core.selector.entitySelectorNegationSample
+     */
+    public fun not(init: NegatedEntitySelectorScope.() -> Unit)
+}