+++ modules/core/src/main/kotlin/io/github/lmliam/kotventure/core/selector/EntitySelectorStateOwner.kt
@@ -1,0 +1,5 @@
+package io.github.lmliam.kotventure.core.selector
+
+internal interface EntitySelectorStateOwner {
+    val selectorState: EntitySelectorState
+}