+++ modules/core/src/main/kotlin/io/github/lmliam/kotventure/core/selector/NegatedEntitySelectorAdapter.kt
@@ -1,0 +1,6 @@
+package io.github.lmliam.kotventure.core.selector
+
+internal class NegatedEntitySelectorAdapter(
+    state: EntitySelectorState,
+) : NegatedSelfEntitySelectorAdapter(state),
+    NegatedEntitySelectorScope