+++ modules/core/src/main/kotlin/io/github/lmliam/kotventure/core/selector/NegatedPlayerEntitySelectorAdapter.kt
@@ -1,0 +1,6 @@
+package io.github.lmliam.kotventure.core.selector
+
+internal class NegatedPlayerEntitySelectorAdapter(
+    state: EntitySelectorState,
+) : NegatedCommonEntitySelectorAdapter(state),
+    NegatedPlayerEntitySelectorScope